+++ src/main/java/net/geforcemods/securitycraft/blockentities/MotionActivatedLightBlockEntity.java
@@ -5,6 +5,8 @@ import java.util.List;
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.api.CustomizableBlockEntity;
 import net.geforcemods.securitycraft.api.Option;
+import net.geforcemods.securitycraft.api.Option.BooleanOption;
+import net.geforcemods.securitycraft.api.Option.DisabledOption;
 import net.geforcemods.securitycraft.api.Option.DoubleOption;
 import net.geforcemods.securitycraft.blocks.MotionActivatedLightBlock;
 import net.geforcemods.securitycraft.entity.Sentry;
@@ -21,6 +23,7 @@ import net.minecraft.world.phys.AABB;
 public class MotionActivatedLightBlockEntity extends CustomizableBlockEntity implements ITickingBlockEntity {
 	private static final int TICKS_BETWEEN_ATTACKS = 5;
 	private DoubleOption searchRadiusOption = new DoubleOption(this::getBlockPos, "searchRadius", 5.0D, 5.0D, 20.0D, 1.0D, true);
+	private DisabledOption disabled = new DisabledOption(false);
 	private int cooldown = TICKS_BETWEEN_ATTACKS;
 
 	public MotionActivatedLightBlockEntity(BlockPos pos, BlockState state) {
@@ -29,7 +32,7 @@ public class MotionActivatedLightBlockEntity extends CustomizableBlockEntity imp
 
 	@Override
 	public void tick(Level level, BlockPos pos, BlockState state) {
-		if (cooldown-- > 0)
+		if (isDisabled() || cooldown-- > 0)
 			return;
 
 		List<LivingEntity> entities = level.getEntitiesOfClass(LivingEntity.class, new AABB(pos).inflate(searchRadiusOption.get()), e -> !EntityUtils.isInvisible(e) && !e.isSpectator() && !(e instanceof Sentry || e instanceof ArmorStand));
@@ -42,6 +45,15 @@ public class MotionActivatedLightBlockEntity extends CustomizableBlockEntity imp
 	}
 
 	@Override
+	public void onOptionChanged(Option<?> option) {
+		//turn off the light when it's disabled
+		if (option.getName().equals("disabled") && ((BooleanOption) option).get()) {
+			if (getBlockState().getValue(MotionActivatedLightBlock.LIT))
+				level.setBlockAndUpdate(worldPosition, getBlockState().setValue(MotionActivatedLightBlock.LIT, false));
+		}
+	}
+
+	@Override
 	public ModuleType[] acceptedModules() {
 		return new ModuleType[] {};
 	}
@@ -49,7 +61,11 @@ public class MotionActivatedLightBlockEntity extends CustomizableBlockEntity imp
 	@Override
 	public Option<?>[] customOptions() {
 		return new Option<?>[] {
-			searchRadiusOption
+				searchRadiusOption, disabled
 		};
+	}
+
+	public boolean isDisabled() {
+		return disabled.get();
 	}
 }