+++ src/main/java/net/geforcemods/securitycraft/blocks/KeyPanelBlock.java
@@ -99,7 +99,9 @@ public class KeyPanelBlock extends OwnableBlock implements SimpleWaterloggedBloc
 		else if (!level.isClientSide) {
 			KeyPanelBlockEntity be = (KeyPanelBlockEntity) level.getBlockEntity(pos);
 
-			if (ModuleUtils.isDenied(be, player)) {
+			if (be.isDisabled())
+				player.displayClientMessage(Utils.localize("gui.securitycraft:scManual.disabled"), true);
+			else if (ModuleUtils.isDenied(be, player)) {
 				if (be.sendsMessages())
 					PlayerUtils.sendMessageToPlayer(player, Utils.localize(getDescriptionId()), Utils.localize("messages.securitycraft:module.onDenylist"), ChatFormatting.RED);
 			}