+++ Changelog.md
@@ -3,7 +3,10 @@
 - New: Items can now be dragged from JEI into the Inventory Scanner's ghost slots
 - New: Item and block tag "securitycraft:reinforced/terracotta"
 - New: Several additions to minecraft's tags, including "minecraft:azalea_grows_on", "minecraft:sand", and "minecraft:terracotta"
+- New: Several blocks now have a new Universal Block Modifier option to disable them
 - Change: Reduced the Rail Mine's explosion size by 25% to make it comparable to other mines' explosions
+- API: Added DisabledOption for having a default option to disable blocks
+- API: Added Option#getKey and Option#getDescriptionKey to easily access the language keys associated with an option
 - Fix: Water doesn't render correctly while a Sonic Security System is recording within the player's view
 - Fix: The reinforced stone pressure plate is not in SecurityCraft's "securitycraft:reinforced/pressure_plates" tag